application/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
requests==2.32.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
